+++ src/events/message/messageUpdate.ts
@@ -15,36 +15,36 @@ export default class MessageUpdateEvent extends Event {
       if (!webhook) return;
       const guild = await bot.utils.getGuildById(newMsg.guild.id);
       if (!guild) return;
-  
+
       const blacklistedWords = guild?.blacklistedwords;
-  
+
       if (!oldMsg.content || !newMsg.content) {
         return;
       }
       if (newMsg.author?.id === bot.user?.id) return;
-  
+
       if (oldMsg.content === newMsg.content) return;
-  
+
       if (blacklistedWords !== null && blacklistedWords[0]) {
         blacklistedWords.forEach((word) => {
           if (newMsg.content.toLowerCase().includes(word.toLowerCase())) {
-            newMsg.delete();
+            newMsg.deletable && newMsg.delete();
             return newMsg
               .reply("You used a bad word the admin has set, therefore your message was deleted!")
               .then((msg) => {
                 setTimeout(() => {
-                  msg.delete();
+                  msg.deletable && msg.delete();
                 }, 5000);
               });
           }
         });
       }
-  
+
       const pOldMsg = oldMsg.content.length > 1024 ? `${oldMsg.content.slice(0, 1010)}...` : oldMsg;
       const PNewMsg = newMsg.content.length > 1024 ? `${newMsg.content.slice(0, 1010)}...` : newMsg;
-  
+
       const messageLink = `https://discord.com/channels/${newMsg.guild.id}/${newMsg.channel.id}/${newMsg.id}`;
-  
+
       const embed = bot.utils
         .baseEmbed(newMsg)
         .setTitle(`Message updated in **${(newMsg.channel as TextChannel).name}**`)
@@ -55,7 +55,7 @@ export default class MessageUpdateEvent extends Event {
         .addField("**New Message**", `${PNewMsg}`)
         .setColor("ORANGE")
         .setTimestamp();
-  
+
       webhook.send(embed);
     } catch (err) {
       bot.utils.sendErrorLog(err, "error");
